requirements.txt
streamlit>=1.32.0
openai>=1.12.0
google-generativeai>=0.4.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
